READ DESCRIPTION!!!!!!!!!!!!!!!!
I updated the code to have the kitchen system work on a seperate thread, ie it runs on the side so that the other code works. This complicates things, but in short, everythinig that goes into the active tickets and past tickets needs to be saved to their respective text files for the moment.